+++ .env.sample
@@ -1,0 +1,12 @@
+# Environment Config
+
+# store your secrets and config variables in here
+# only invited collaborators will be able to see your .env values
+
+# reference these in your code with process.env.SECRET
+
+SLACK_BOT_TOKEN=<From https://api.slack.com/apps/APP_ID/oauth>
+SLACK_VERIFICATION_TOKEN=<From https://api.slack.com/apps/APP_ID>
+APIAI_TOKEN=<From API.AI console->Settings->Developer access token>
+
+# note: .env is a shell file so there can't be spaces around =
