+++ .gitignore
@@ -2,3 +2,5 @@ node_modules
 .DS_Store
 .vscode
 config.json
+.env
+
