--- config.default.json
@@ -1,5 +1,0 @@
-{
-  "SLACK_TOKEN": "<YOUR_SLACK_TOKEN>",
-  "APIAI_TOKEN": "<YOUR_APIAI_TOKEN>",
-  "VERIFICATION_TOKEN": "<YOUR_VERIFICATION_TOKEN>"
-}
