--- keanubot_apiai_agent/agent.json
@@ -1,24 +1,0 @@
-{
-  "description": "",
-  "language": "en",
-  "enabledDomainFeatures": [],
-  "googleAssistant": {
-    "googleAssistantCompatible": false,
-    "welcomeIntentSignInRequired": false,
-    "startIntents": [],
-    "endIntentIds": [],
-    "oAuthLinking": {
-      "required": false,
-      "grantType": "AUTH_CODE_GRANT"
-    },
-    "voiceType": "MALE_1"
-  },
-  "defaultTimezone": "",
-  "webhook": {
-    "available": false,
-    "useForDomains": false
-  },
-  "isPrivate": false,
-  "customClassifierMode": "use.after",
-  "mlMinConfidence": 0.2
-}
