--- keanubot_apiai_agent/intents/Excellent.json
@@ -1,49 +1,0 @@
-{
-  "userSays": [
-    {
-      "id": "a6451b25-fae1-4151-b6cf-e58eaf66acad",
-      "data": [
-        {
-          "text": "congrats"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    },
-    {
-      "id": "3bd9f310-12db-4f5c-b1f7-e986ea92770a",
-      "data": [
-        {
-          "text": "perfect"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    }
-  ],
-  "id": "3265b1a0-16c6-4f78-bfc7-ea365d259758",
-  "name": "Excellent",
-  "auto": true,
-  "contexts": [],
-  "responses": [
-    {
-      "resetContexts": false,
-      "affectedContexts": [],
-      "parameters": [],
-      "messages": [
-        {
-          "type": 0,
-          "speech": [
-            "Most Excellent, http://cdn.bloody-disgusting.com/wp-content/uploads/2016/06/bill-and-ted-s-excellent-adventure_55029933-3000x2005-3000x2005.jpeg",
-            "Most Excellent, https://www.joeforrest.com/wp-content/uploads/2016/09/Be-Excellent-to-one-another.jpg"
-          ]
-        }
-      ]
-    }
-  ],
-  "priority": 500000,
-  "webhookUsed": false,
-  "webhookForSlotFilling": false,
-  "fallbackIntent": false,
-  "events": []
-}
