--- keanubot_apiai_agent/intents/GetOnline.json
@@ -1,46 +1,0 @@
-{
-  "userSays": [
-    {
-      "id": "d4caa486-fc3a-43ac-af8e-17848af20db3",
-      "data": [
-        {
-          "text": "what do you need"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    },
-    {
-      "id": "3e6d4559-5e35-4100-aa8c-c89a764c95b4",
-      "data": [
-        {
-          "text": "I need"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    }
-  ],
-  "id": "5fa2f6b9-548a-4705-af4b-9629ca7d5fbd",
-  "name": "GetOnline",
-  "auto": true,
-  "contexts": [],
-  "responses": [
-    {
-      "resetContexts": false,
-      "affectedContexts": [],
-      "parameters": [],
-      "messages": [
-        {
-          "type": 0,
-          "speech": "I need to get online........, https://68.media.tumblr.com/1086c071cb2d7a9b034de2c055be4019/tumblr_mq2e12q1hp1szspk7o1_500.gif"
-        }
-      ]
-    }
-  ],
-  "priority": 500000,
-  "webhookUsed": false,
-  "webhookForSlotFilling": false,
-  "fallbackIntent": false,
-  "events": []
-}
