--- keanubot_apiai_agent/intents/Guns.json
@@ -1,76 +1,0 @@
-{
-  "userSays": [
-    {
-      "id": "bd1cfb92-7dfa-49d6-8778-ce3c9d908978",
-      "data": [
-        {
-          "text": "need anything ?"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    },
-    {
-      "id": "081b0166-c573-4381-b8af-960a0bd5fbf9",
-      "data": [
-        {
-          "text": "tell me what you need"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    },
-    {
-      "id": "16740deb-eee5-497c-8cbf-7048d1b64d0a",
-      "data": [
-        {
-          "text": "can I get you something"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    },
-    {
-      "id": "02c34982-9934-4db4-bcd7-f7eaa7324db7",
-      "data": [
-        {
-          "text": "what are your requirements"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    },
-    {
-      "id": "cdee215c-92d5-4ff1-bf05-8b5cf4989066",
-      "data": [
-        {
-          "text": "what do you need"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    }
-  ],
-  "id": "12c81a0a-d051-4a25-bd6f-37a1e88d7466",
-  "name": "Guns",
-  "auto": true,
-  "contexts": [],
-  "responses": [
-    {
-      "resetContexts": false,
-      "affectedContexts": [],
-      "parameters": [],
-      "messages": [
-        {
-          "type": 0,
-          "speech": "I need guns. Lots of guns. , http://m.memegen.com/we34pq.jpg"
-        }
-      ]
-    }
-  ],
-  "priority": 500000,
-  "webhookUsed": false,
-  "webhookForSlotFilling": false,
-  "fallbackIntent": false,
-  "events": []
-}
