--- keanubot_apiai_agent/intents/KungFu.json
@@ -1,66 +1,0 @@
-{
-  "userSays": [
-    {
-      "id": "f9307d13-4ba2-4c69-a48f-5fbd0da2917f",
-      "data": [
-        {
-          "text": "kung fu ?"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    },
-    {
-      "id": "6237f74a-f989-45db-8cd5-cada647e5a03",
-      "data": [
-        {
-          "text": "what else do you know"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    },
-    {
-      "id": "5d422f65-1c5a-40e3-bae8-a37a6cbe2411",
-      "data": [
-        {
-          "text": "do you know"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    },
-    {
-      "id": "7dd6aa9e-74c2-46e5-b3bf-0d80dbdc3cb5",
-      "data": [
-        {
-          "text": "know how to"
-        }
-      ],
-      "isTemplate": false,
-      "count": 1
-    }
-  ],
-  "id": "05984c29-a7e0-4892-9e60-3452190f5955",
-  "name": "KungFu",
-  "auto": true,
-  "contexts": [],
-  "responses": [
-    {
-      "resetContexts": false,
-      "affectedContexts": [],
-      "parameters": [],
-      "messages": [
-        {
-          "type": 0,
-          "speech": "I know kung fu, http://kungfukingdom.com/wp-content/uploads/2016/08/The-Matrix-Kung-Fu-Kingdom-770x472.jpg"
-        }
-      ]
-    }
-  ],
-  "priority": 500000,
-  "webhookUsed": false,
-  "webhookForSlotFilling": false,
-  "fallbackIntent": false,
-  "events": []
-}
