--- keanubot_apiai_agent/intents/My name is NEO.json
@@ -1,36 +1,0 @@
-{
-  "userSays": [
-    {
-      "id": "d9531737-8be6-4099-8730-ad08f2db5120",
-      "data": [
-        {
-          "text": "My name is"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    }
-  ],
-  "id": "14759782-8357-460a-911d-f46dfba022f0",
-  "name": "My name is NEO",
-  "auto": true,
-  "contexts": [],
-  "responses": [
-    {
-      "resetContexts": false,
-      "affectedContexts": [],
-      "parameters": [],
-      "messages": [
-        {
-          "type": 0,
-          "speech": "My name ...... is NEO ,  http://1.bp.blogspot.com/-3S6-8oHs7l8/T6vxvqlLuXI/AAAAAAAABkU/qGp-BkfXdiU/s1600/mynameisneo.jpg"
-        }
-      ]
-    }
-  ],
-  "priority": 500000,
-  "webhookUsed": false,
-  "webhookForSlotFilling": false,
-  "fallbackIntent": false,
-  "events": []
-}
