--- keanubot_apiai_agent/intents/OldVSNew.json
@@ -1,46 +1,0 @@
-{
-  "userSays": [
-    {
-      "id": "cd26e0c3-da55-4c9f-b31d-cab315315a49",
-      "data": [
-        {
-          "text": "I\u0027m getting Old"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    },
-    {
-      "id": "335110cf-0f85-4c06-932f-be47ad770e18",
-      "data": [
-        {
-          "text": "I\u0027m feeling old"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    }
-  ],
-  "id": "cd282c5f-f67f-4f98-8ee3-94217aab1f26",
-  "name": "OldVSNew",
-  "auto": true,
-  "contexts": [],
-  "responses": [
-    {
-      "resetContexts": false,
-      "affectedContexts": [],
-      "parameters": [],
-      "messages": [
-        {
-          "type": 0,
-          "speech": "So am I, http://i.imgur.com/2bbCh.jpg"
-        }
-      ]
-    }
-  ],
-  "priority": 500000,
-  "webhookUsed": false,
-  "webhookForSlotFilling": false,
-  "fallbackIntent": false,
-  "events": []
-}
