--- keanubot_apiai_agent/intents/Speed.json
@@ -1,69 +1,0 @@
-{
-  "userSays": [
-    {
-      "id": "961d18a0-8533-4656-bedf-bc32e21d270e",
-      "data": [
-        {
-          "text": "I\u0027m on bus "
-        },
-        {
-          "text": "23",
-          "alias": "number",
-          "meta": "@sys.number",
-          "userDefined": false
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    },
-    {
-      "id": "829b0181-134f-4161-ab81-73df35923d19",
-      "data": [
-        {
-          "text": "Get on the bus"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    },
-    {
-      "id": "dd092af5-2787-4fe6-8d15-974570b97d2c",
-      "data": [
-        {
-          "text": "Did you catch the bus"
-        }
-      ],
-      "isTemplate": false,
-      "count": 0
-    }
-  ],
-  "id": "45c1e025-4fdb-4a7c-8011-e5c34e73208f",
-  "name": "Speed",
-  "auto": true,
-  "contexts": [],
-  "responses": [
-    {
-      "resetContexts": false,
-      "affectedContexts": [],
-      "parameters": [
-        {
-          "dataType": "@sys.number",
-          "name": "number",
-          "value": "$number",
-          "isList": false
-        }
-      ],
-      "messages": [
-        {
-          "type": 0,
-          "speech": "you mean this bus ?,http://i59.tinypic.com/5n6h3a.jpg"
-        }
-      ]
-    }
-  ],
-  "priority": 500000,
-  "webhookUsed": false,
-  "webhookForSlotFilling": false,
-  "fallbackIntent": false,
-  "events": []
-}
